requirements.txt
opencv-python
numpy
PyYAML
pandas